Introduce -standard and -standard-commands artifacts
Fixes #43
