+++ .gitignore
@@ -8,3 +8,4 @@ target/
 *.iml
 *.ipr
 *.iws
+.DS_Store
