+++ pom.xml
@@ -21,6 +21,8 @@
 	
 	<modules>
 		<module>spring-shell2-core</module>
+		<module>spring-shell2-standard</module>
+		<module>spring-shell2-standard-commands</module>
 		<module>spring-shell2-jcommander-adapter</module>
 		<module>spring-shell2-samples</module>
 		<module>spring-shell2-shell1-adapter</module>
@@ -31,6 +33,16 @@
 			<dependency>
 				<groupId>org.springframework.shell</groupId>
 				<artifactId>spring-shell2-core</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.shell</groupId>
+				<artifactId>spring-shell2-standard</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.shell</groupId>
+				<artifactId>spring-shell2-standard-commands</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
