+++ spring-shell2-samples/pom.xml
@@ -32,6 +32,15 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.shell</groupId>
+			<artifactId>spring-shell2-standard</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.shell</groupId>
+			<artifactId>spring-shell2-standard-commands</artifactId>
+			<scope>runtime</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.shell</groupId>
 			<artifactId>spring-shell2-shell1-adapter</artifactId>
 		</dependency>
 		<dependency>
