+++ spring-shell2-standard-commands/src/test/java/org/springframework/shell2/standard/commands/HelpTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.shell2.commands;
+package org.springframework.shell2.standard.commands;
 
 import java.io.IOException;
 import java.io.InputStream;
