+++ spring-shell2-standard-commands/src/test/resources/org/springframework/shell2/standard/commands/HelpTest-testCommandList.txt
