+++ spring-shell2-standard/src/main/java/org/springframework/shell2/standard/ValueProviderSupport.java
