+++ spring-shell2-standard/src/test/java/org/springframework/shell2/standard/CommandValueProviderTest.java
